fix.修复一些warning

+++ src/MeoAssistant/InfrastAbstractTask.cpp
@@ -16,12 +16,12 @@ asst::InfrastAbstractTask::InfrastAbstractTask(AsstCallback callback, void* call
     : AbstractTask(callback, callback_arg)
 {
     if (m_face_hash_thres == 0) {
-        m_face_hash_thres = std::dynamic_pointer_cast<MatchTaskInfo>(
-            Task.get("InfrastOperFaceHash"))->templ_threshold;
+        m_face_hash_thres = static_cast<int>(std::dynamic_pointer_cast<MatchTaskInfo>(
+            Task.get("InfrastOperFaceHash"))->templ_threshold);
     }
     if (m_name_hash_thres == 0) {
-        m_name_hash_thres = std::dynamic_pointer_cast<MatchTaskInfo>(
-            Task.get("InfrastOperNameHash"))->templ_threshold;
+        m_name_hash_thres = static_cast<int>(std::dynamic_pointer_cast<MatchTaskInfo>(
+            Task.get("InfrastOperNameHash"))->templ_threshold);
     }
 }
